refactored loginSlice, moved auth check to RootLayout
