+++ src/components/Bar/Bar.js
@@ -1,14 +1,8 @@
-import React, { useCallback } from "react";
+import React from "react";
 import AppBar from "@mui/material/AppBar";
 import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
 import { useNavigate } from "react-router-dom";
-import { useDispatch } from "react-redux";
-import { useEffect } from "react";
-import { loginActions } from "../../store/loginSlice";
-import { onAuthStateChanged } from "firebase/auth";
-import { getCurrentUser } from "../../api/FirebaseAuthApi";
-import { auth } from "../../configuration/firebase/FirebaseCommon";
 import { Container, ThemeProvider } from "@mui/material";
 import { Logo } from "../Logo/LogoIcon";
 import { lightPurpleTheme } from "../../themes/purpleTheme";
@@ -16,30 +10,6 @@ import { NavigationPanel } from "./NavigationPanel";
 
 const Bar = () => {
   const navigate = useNavigate();
-  const dispatch = useDispatch();
-
-  const handleAdminRigths = useCallback(async () => {
-    const token = await getCurrentUser().getIdTokenResult();
-    dispatch(loginActions.setAdmin(token.claims.role === 'admin'));
-  }, [dispatch]);
-
-  const authStateObserver = useCallback(
-    (user) => {
-      if (user) {
-        dispatch(loginActions.setLoggedIn(true));
-        handleAdminRigths();
-        dispatch(loginActions.setEmail(user.email));
-        dispatch(loginActions.setImageUrl(user.photoURL));
-        dispatch(loginActions.setUserId(user.uid));
-      } else {
-      }
-    },
-    [dispatch, handleAdminRigths]
-  );
-
-  useEffect(() => {
-    onAuthStateChanged(auth, authStateObserver);
-  }, [authStateObserver]);
 
   const goToMainPageHandler = () => {
     navigate("/");
