+++ src/components/UserSettingsForm/UserSettingsForm.js
@@ -11,7 +11,7 @@ import Stack from "@mui/material/Stack"
 import { styled, useMediaQuery, useTheme } from "@mui/material";
 import { uploadUserPhoto } from "../../api/FirestoreApi";
 import { updatePhoto } from "../../api/FirebaseAuthApi";
-import { loginActions } from "../../store/loginSlice";
+import { setImageUrl } from "../../store/loginSlice";
 import { DropzoneMobile } from "../Dropzone/DropzoneMobile";
 import { Dropzone } from "../Dropzone/Dropzone";
 import { PrimaryButton } from "../PrimaryButton/PrimaryButton";
@@ -47,7 +47,7 @@ const UserSettingsForm = () => {
     }
     try {
       updatePhoto(userId, url);
-      dispatch(loginActions.setImageUrl(url));
+      dispatch(setImageUrl(url));
 
       navigate("/", { state: { message: "Photo was uploaded!" } });
     } catch (error) {
